Sop_impoter
This simple script for Houdini allows you to automate the process of creating sopimport and materials inside the /stage/ directory. It searches for the selected geometric nodes and creates for each one a sopimport and a materialiallibrary which are linked to each other. It also creates a scene import node and places all the nodes in a specific order on the graphics panel.

Note: currently, the script does not support pulling the geometry path from solaris to the library material parameter called geometry path. It is also planned to add file connection in the future.